backend/neurofin_backend/services/risk/requirements.txt
flask==2.3.2
pymongo==4.4.0
scikit-learn==1.3.2
numpy==1.26.0
python-dotenv==1.0.0
